delete unused programs. bif_ns and fix is merged to bif.
