+++ bif/newton.cpp
@@ -1,24 +1,43 @@
 #include "newton.hpp"
-#include "ds_func.hpp"
 #include "dynamical_system.hpp"
+#include <fstream>
+#include <iomanip>
 
 void newton(dynamical_system &ds) {
-  Eigen::VectorXd vp(ds.xdim + 1);
-  // vp << ds.x0 << ds.params(ds.var_param);
-  vp(Eigen::seqN(0, ds.xdim)) = ds.x0;
-  vp(ds.xdim) = ds.params(ds.var_param);
-  Eigen::VectorXd vn(ds.xdim + 1);
-  Eigen::VectorXd F(ds.xdim + 1);
-  Eigen::MatrixXd J(ds.xdim + 1, ds.xdim + 1);
+  unsigned int target_dim;
+  if (!ds.fix_mode) {
+    target_dim = ds.xdim + 2;
+  } else {
+    target_dim = ds.xdim;
+  }
+  Eigen::VectorXd vp(target_dim);
+  Eigen::VectorXd vn(target_dim);
+  Eigen::VectorXd F(target_dim);
+  Eigen::MatrixXd J(target_dim, target_dim);
   double norm;
+  Eigen::IOFormat Comma(Eigen::StreamPrecision, 0, ", ", "\n", "[", "]");
+  std::cout << std::fixed << std::setprecision(16);
+  std::vector<Eigen::VectorXd> bifset(ds.inc_iter,
+                                      Eigen::VectorXd::Zero(ds.p.size()));
+  bool exit_flag = false;
+
+  vp(Eigen::seqN(0, ds.xdim)) = ds.x0;
+  if (!ds.fix_mode) {
+    vp(ds.xdim) = ds.p(ds.var_param);
+    vp(ds.xdim + 1) = ds.theta;
+  }
 
   for (int p = 0; p < ds.inc_iter; p++) {
+    if (exit_flag) {
+      bifset.erase(bifset.begin() + p, bifset.end());
+      ds.inc_iter = p - 1;
+      break;
+    }
     auto start = std::chrono::system_clock::now();
     for (int i = 0; i < ds.max_iter; i++) {
-      store_state(vp, ds);
-
-      F = func_newton(ds);
-      J = jac_newton(ds);
+      auto FJ = ds.newton_FJ(vp);
+      F = std::get<0>(FJ);
+      J = std::get<1>(FJ);
       vn = Eigen::ColPivHouseholderQR<Eigen::MatrixXd>(J).solve(-F) + vp;
 
       norm = (vn - vp).norm();
@@ -29,12 +48,15 @@ void newton(dynamical_system &ds) {
             std::chrono::duration_cast<std::chrono::milliseconds>(dur).count();
         std::cout << "**************************************************"
                   << std::endl;
+        std::cout << std::setprecision(16);
         std::cout << p << " : converged (iter = " << i + 1 << ", ";
         std::cout << "time = " << msec << "[msec])" << std::endl;
-        std::cout << "params : " << ds.params.transpose() << std::endl;
+        std::cout << "params : " << ds.p.transpose().format(Comma) << std::endl;
         std::cout << "x0     : "
-                  << vn(Eigen::seqN(0, ds.xdim)).real().transpose()
+                  << vn(Eigen::seqN(0, ds.xdim)).transpose().format(Comma)
                   << std::endl;
+        std::cout << "theta  : " << ds.theta << std::endl;
+        std::cout << std::setprecision(4);
         std::cout << "(Re(μ), Im(μ)), abs(μ), arg(μ) :" << std::endl;
         for (int k = 0; k < ds.xdim; k++) {
           std::cout << ds.eigvals(k) << ", ";
@@ -43,22 +65,37 @@ void newton(dynamical_system &ds) {
         }
         std::cout << "**************************************************"
                   << std::endl;
+        bifset[p] = ds.p;
         vp = vn;
-        ds.params(ds.var_param) = vn(ds.xdim);
         break;
       } else if (norm >= ds.explode) {
         std::cerr << "explode (iter = " << i + 1 << ")" << std::endl;
-        exit(1);
+        exit_flag = true;
+        break;
       }
-
       if (i >= ds.max_iter - 1) {
         std::cerr << "iter over (iter = " << i + 1 << ")" << std::endl;
-        exit(1);
+        exit_flag = true;
+        break;
       }
-
       vp = vn;
-      ds.params(ds.var_param) = vn(ds.xdim);
     }
-    ds.params[ds.inc_param] += ds.delta_inc;
+    ds.p[ds.inc_param] += ds.delta_inc;
+  }
+
+  // set last state
+  ds.x0 = vn(Eigen::seqN(0, ds.xdim));
+  // output
+  if (!ds.fix_mode) {
+    ds.p(ds.var_param) = vn(ds.xdim);
+    ds.theta = vn(ds.xdim + 1);
+    Eigen::IOFormat Out(Eigen::FullPrecision, 0, " ", "\n", " ", " ");
+    std::ofstream f;
+    f.open(ds.out_path, std::ios::out);
+    f << std::fixed;
+    for (int i = 0; i < ds.inc_iter; i++) {
+      f << bifset[i].transpose().format(Out) << std::endl;
+    }
+    f.close();
   }
 }
