+++ bif/sys_func.cpp
