--- bif/y3PD2.json
@@ -1,46 +1,0 @@
-{
-    "_comment": "This is a samle setting file",
-    "xrange": [
-        -3.0,
-        3
-    ],
-    "yrange": [
-        -3.0,
-        3.0
-    ],
-    "alpha": 0.25,
-    "break": 16,
-    "x0": [
-        -0.7562134639782205,
-        0.8678720384681493
-    ],
-    "params": [
-        4.6640000000000086,
-        1,
-        0.005,
-        1.0,
-        3
-    ],
-    "dparams": [
-        0.01,
-        0.1,
-        0.1,
-        0.1
-    ],
-    "period": 2,
-    "sigma": -1,
-    "omega": 0,
-    "inc_param": 3,
-    "var_param": 0,
-    "delta_inc": -0.0001,
-    "inc_iter": 100,
-    "max_iter": 32,
-    "numerical_diff": false,
-    "dif_strip": 1e-08,
-    "eps": 1e-08,
-    "explode": 1000,
-    "func": [
-        "data.dict['params'][3] * x[1]",
-        "data.dict['params'][0] * x[1]**3 * (x[1]**2 - data.dict['params'][1]**2) + data.dict['params'][2] * x[0]"
-    ]
-}
