--- bif_ns/newton.hpp
@@ -1,11 +1,0 @@
-#ifndef NEWTON_HPP_
-#define NEWTON_HPP_
-
-#include "sys_common.hpp"
-#include <chrono>
-
-class dynamical_system;
-
-void newton(dynamical_system &ds);
-
-#endif
